+++ CHANGES.rst
@@ -4,7 +4,8 @@ Changelog
 6.1.10 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Added description to PDC fields
+  [pnicolli]
 
 
 6.1.9 (2024-01-11)