+++ CHANGES.rst
@@ -6,6 +6,8 @@ Changelog
 
 - Added description to PDC fields
   [pnicolli]
+  Added upgrade step to update PDC fields description
+  [lucabel]
 
 
 6.1.9 (2024-01-11)