+++ CHANGES.rst
@@ -11,6 +11,7 @@ Changelog
 - Added new widget for event luoghi_correlati
   [pnicolli]
 
+
 6.1.9 (2024-01-11)
 ------------------